[lldb-test] Help dotest to find the directory where clang is built.
This is really not ideal, but needed. Upstream does the same,
I'm going to probably have a discussion there to avoid this
depedency but in the meanwhile, we're trying to get the
bot green again. I think this has still some missing cases,
as swift adds a `-Asan` suffix when building with asan.